requirements.txt
Click==7.0
requests==2.22.0
gunicorn==20.0.4
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.17
psycopg2==2.9.3
Werkzeug==0.16.0
python-dotenv==0.20.0
